# Conflicts:
#	source/Game/game.c

+++ source/Game/ball.h
@@ -1,19 +1,11 @@
-//typedef struct Ball{
-//	int x=0;
-//	int y=0;
-//	int size=25;
-//	int widthSize=50;
-//	int heightSize=49;
-//};
-//void drawBall(struct Ball* self);
 
-//typedef struct Ball ;
+// struct Ball ;
 struct Ball {
-	double x;
-	double y;
+	float x;
+	float y;
 	int radius;
-	double speed;
-	double angle;
+	float speed;
+	float angle;
 };
 void draw_ball(struct Ball *self);
 void move_ball(struct Ball *self, int block_layout[][2]);
